# src/docs/openapi.dev.json
{
  "openapi": "3.0.3",
  "info": { "title": "Business API (Dev)", "version": "1.0.0-dev", "description": "Full specification including Dev/Debug endpoints" },
  "servers": [ { "url": "http://localhost:3000" } ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "apiKeyAuth": { "type": "apiKey", "in": "header", "name": "Authorization", "description": "Use 'Authorization: Token ck_...' for customer API key auth" }
    },
    "schemas": {
      "OkWrapper": { "type": "object", "properties": { "success": { "type": "boolean", "example": true }, "data": { "type": "object" } } },
      "ErrorResponse": { "type": "object", "properties": { "success": { "type": "boolean", "example": false }, "error": { "type": "object", "properties": { "code": { "type": "string" }, "message": { "type": "string" } }, "required": ["code", "message"] } }, "required": ["success", "error"] },
      "SignupRequest": { "type": "object", "properties": { "email": { "type": "string", "format": "email" }, "password": { "type": "string", "minLength": 8 }, "company": { "type": "string" }, "plan": { "type": "string", "enum": ["basic", "pro"] } }, "required": ["email", "password"] },
...
    }
  },
  "paths": {
...
    "/api/v1/customer/auth/signup": { "post": { "tags": ["Auth","Customer"], "summary": "Customer signup (JWT)", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SignupRequest" } } } }, "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OkWrapper" } } } }, "400": { "description": "Weak password / bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }, "409": { "description": "Email exists", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } } } },
    "/api/v1/customer/auth/login": { "post": { "tags": ["Auth","Customer"], "summary": "Customer login (JWT)", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "email": { "type": "string", "format": "email" }, "password": { "type": "string" } }, "required": ["email","password"] } } } }, "responses": { "200": { "description": "Logged in", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OkWrapper" } } } }, "401": { "description": "Invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } } } },
...
  }
}